Add files via upload
This function take a sample JSON block and identifies words and phrases that could need to be flagged from a custom list or other possible use cases. For the sake of this example code, I had just finished watching the movie 'The Bourne Identity' so I was having some fun with the naming conventions. Please note these are NOT official in any shape or form.