Add files via upload
Function that will identify words and phrases that could be marked for redaction - result is the phrase along with position.  For the sake of this example code, I had just finished watching the movie 'The Bourne Identity' so I was having some fun with the naming conventions. Please note these are NOT official in any shape or form.